[s8s] Update tunnel tests, move to `tunnel/`
