+++ src/commands/synthetics/__tests__/tunnel/websocket.test.ts
@@ -1,10 +1,10 @@
-import * as ciUtils from '../../../helpers/utils'
+import * as ciUtils from '../../../../helpers/utils'
 
-import {DEFAULT_POLLING_TIMEOUT, RunTestCommand} from '../command'
-import {ExecutionRule} from '../interfaces'
-import * as utils from '../utils'
+import {DEFAULT_POLLING_TIMEOUT, RunTestCommand} from '../../command'
+import {ExecutionRule} from '../../interfaces'
+import * as utils from '../../utils'
 
-import {getSyntheticsProxy} from './fixtures'
+import {getSyntheticsProxy} from '../fixtures'
 
 describe('Proxy configuration', () => {
   let initialHttpProxyEnv: string | undefined
